+++ Library/HK32F030Mxx_Library_V1.1.6/HK32F030M_Project/tests/mocks/mock_status_leds_hw.c
@@ -40,7 +40,7 @@ void status_leds_hw_refresh()
     function_called();
 }
 
-void status_leds_hw_set_brightness(float brightness)
+void status_leds_hw_set_brightness(uint8_t brightness)
 {
     check_expected(brightness);
 }
